+++ Passepartout-iOS/Global/SwiftGen+Scenes.swift
@@ -38,6 +38,8 @@ internal enum StoryboardScene {
     internal static let storyboardName = "Shortcuts"
 
     internal static let initialScene = InitialSceneType<UIKit.UINavigationController>(storyboard: Shortcuts.self)
+
+    internal static let shortcutsViewController = SceneType<UIKit.UINavigationController>(storyboard: Shortcuts.self, identifier: "ShortcutsViewController")
   }
 }
 // swiftlint:enable explicit_type_interface identifier_name line_length type_body_length type_name