+++ Passepartout-iOS/Scenes/Organizer/ShortcutsViewController.swift
@@ -2,7 +2,7 @@
 //  ShortcutsViewController.swift
 //  Passepartout-iOS
 //
-//  Created by Davide De Rosa on 3/18/19.
+//  Created by Davide De Rosa on 3/27/19.
 //  Copyright (c) 2019 Davide De Rosa. All rights reserved.
 //
 //  https://github.com/passepartoutvpn
@@ -24,28 +24,66 @@
 //
 
 import UIKit
+import Intents
 import IntentsUI
 import Passepartout_Core
 
-class ShortcutsViewController: UITableViewController, TableModelHost {
+@available(iOS 12, *)
+protocol ShortcutsIntentDelegate: class {
+    func shortcutsDidSelectIntent(intent: INIntent)
+}
 
+@available(iOS 12, *)
+private struct ShortcutWrapper: Comparable {
+    let phrase: String
+
+    let intentDescription: String?
+    
+    let original: INVoiceShortcut
+
+    static func from(_ vs: INVoiceShortcut) -> ShortcutWrapper {
+        return ShortcutWrapper(
+            phrase: vs.invocationPhrase,
+            intentDescription: vs.shortcut.intent?.intentDescription,
+            original: vs
+        )
+    }
+    
+    // MARK: Equatable
+    
+    static func ==(lhs: ShortcutWrapper, rhs: ShortcutWrapper) -> Bool {
+        return lhs.phrase == rhs.phrase
+    }
+
+    // MARK: Comparable
+    
+    static func <(lhs: ShortcutWrapper, rhs: ShortcutWrapper) -> Bool {
+        return lhs.phrase < rhs.phrase
+    }
+}
+
+@available(iOS 12, *)
+class ShortcutsViewController: UITableViewController, INUIAddVoiceShortcutViewControllerDelegate, INUIEditVoiceShortcutViewControllerDelegate, ShortcutsIntentDelegate, TableModelHost {
+    private var wrappers: [ShortcutWrapper]?
+    
+    private var pendingShortcut: INShortcut?
+    
+    private var editedIndexPath: IndexPath?
+    
     // MARK: TableModel
     
     let model: TableModel<SectionType, RowType> = {
         let model: TableModel<SectionType, RowType> = TableModel()
-        model.add(.vpn)
-        model.add(.wifi)
-        model.add(.cellular)
-        model.set([.connect, .enableVPN, .disableVPN], in: .vpn)
-        model.set([.trustCurrentWiFi, .untrustCurrentWiFi], in: .wifi)
-        model.set([.trustCellular, .untrustCellular], in: .cellular)
-        model.setHeader(L10n.Shortcuts.Sections.Vpn.header, for: .vpn)
-        model.setHeader(L10n.Shortcuts.Sections.Wifi.header, for: .wifi)
-        model.setHeader(L10n.Shortcuts.Sections.Cellular.header, for: .cellular)
+        model.add(.all)
+        model.setHeader(L10n.Shortcuts.Edit.Sections.All.header, for: .all)
+        model.set([], in: .all)
         return model
     }()
     
     func reloadModel() {
+        var rows = [RowType](repeating: .shortcut, count: wrappers?.count ?? 0)
+        rows.append(.addShortcut)
+        model.set(rows, in: .all)
     }
     
     // MARK: UIViewController
@@ -54,32 +92,76 @@ class ShortcutsViewController: UITableViewController, TableModelHost {
         super.viewDidLoad()
         
         title = L10n.Organizer.Cells.SiriShortcuts.caption
+
+        INVoiceShortcutCenter.shared.getAllVoiceShortcuts { [weak self] (shortcuts, error) in
+            DispatchQueue.main.async {
+                guard let shortcuts = shortcuts else {
+                    self?.handleShortcutsFetchError(error)
+                    return
+                }
+                self?.handleShortcuts(shortcuts)
+            }
+        }
     }
-}
+    
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
 
-extension ShortcutsViewController {
+        guard let shortcut = pendingShortcut else {
+            return
+        }
+        pendingShortcut = nil
+        let vc = INUIAddVoiceShortcutViewController(shortcut: shortcut)
+        vc.delegate = self
+        present(vc, animated: true, completion: nil)
+    }
+
+    // MARK: Actions
+    
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        if let vc = segue.destination as? ShortcutsAddViewController {
+            vc.delegate = self
+        }
+    }
+    
+    private func addShortcut() {
+        perform(segue: StoryboardSegue.Shortcuts.shortcutAddSegueIdentifier)
+    }
+
+    private func handleShortcutsFetchError(_ error: Error?) {
+        
+        // TODO: really show it?
+//        let alert = Macros.alert(
+//            title,
+//            L10n.Shortcuts.Edit.message(error?.localizedDescription ?? "")
+//        )
+//        alert.addCancelAction(L10n.Global.ok) {
+//            self.close()
+//        }
+//        present(alert, animated: true, completion: nil)
+    }
+
+    private func handleShortcuts(_ shortcuts: [INVoiceShortcut]) {
+        wrappers = shortcuts.map { ShortcutWrapper.from($0) }
+        wrappers?.sort()
+        reloadModel()
+        tableView.reloadData()
+    }
+
+    @IBAction private func close() {
+        dismiss(animated: true, completion: nil)
+    }
+
+    // MARK: UITableViewController
+    
     enum SectionType {
-        case vpn
-
-        case wifi
-
-        case cellular
+        case all
     }
     
     enum RowType {
-        case connect // host or provider+location
-        
-        case enableVPN
-        
-        case disableVPN
-        
-        case trustCurrentWiFi
-        
-        case untrustCurrentWiFi
-        
-        case trustCellular
-        
-        case untrustCellular
+        case shortcut
+
+        case addShortcut
     }
     
     override func numberOfSections(in tableView: UITableView) -> Int {
@@ -97,127 +179,95 @@ extension ShortcutsViewController {
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = Cells.setting.dequeue(from: tableView, for: indexPath)
         switch model.row(at: indexPath) {
-        case .connect:
-            cell.leftText = L10n.Shortcuts.Cells.Connect.caption
-            
-        case .enableVPN:
-            cell.leftText = L10n.Shortcuts.Cells.EnableVpn.caption
-            
-        case .disableVPN:
-            cell.leftText = L10n.Shortcuts.Cells.DisableVpn.caption
-            
-        case .trustCurrentWiFi:
-            cell.leftText = L10n.Shortcuts.Cells.TrustCurrentWifi.caption
-            
-        case .untrustCurrentWiFi:
-            cell.leftText = L10n.Shortcuts.Cells.UntrustCurrentWifi.caption
-            
-        case .trustCellular:
-            cell.leftText = L10n.Shortcuts.Cells.TrustCellular.caption
-            
-        case .untrustCellular:
-            cell.leftText = L10n.Shortcuts.Cells.UntrustCellular.caption
+        case .shortcut:
+            guard let wrapper = wrappers?[indexPath.row] else {
+                break
+            }
+            cell.apply(Theme.current)
+            cell.leftText = wrapper.phrase
+            cell.rightText = wrapper.intentDescription
+
+        case .addShortcut:
+            cell.applyAction(Theme.current)
+            cell.leftText = L10n.Shortcuts.Edit.Cells.AddShortcut.caption
+            cell.accessoryType = .none
+            cell.isTappable = true
         }
-        cell.apply(Theme.current)
         return cell
     }
     
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        guard #available(iOS 12, *) else {
-            return
-        }
         switch model.row(at: indexPath) {
-        case .connect:
-            addConnect()
-            
-        case .enableVPN:
-            addEnable()
-            
-        case .disableVPN:
-            addDisable()
-            
-        case .trustCurrentWiFi:
-            addTrustWiFi()
-            
-        case .untrustCurrentWiFi:
-            addUntrustWiFi()
-            
-        case .trustCellular:
-            addTrustCellular()
-            
-        case .untrustCellular:
-            addUntrustCellular()
-        }
-    }
-}
-
-// MARK: Actions
-
-@available(iOS 12, *)
-extension ShortcutsViewController {
-    private func addConnect() {
-        guard TransientStore.shared.service.hasProfiles() else {
-            let alert = Macros.alert(
-                L10n.Shortcuts.Cells.Connect.caption,
-                L10n.Shortcuts.Alerts.NoProfiles.message
-            )
-            alert.addAction(L10n.Global.ok) {
-                if let ip = self.tableView.indexPathForSelectedRow {
-                    self.tableView.deselectRow(at: ip, animated: true)
-                }
+        case .shortcut:
+            guard let wrapper = wrappers?[indexPath.row] else {
+                break
             }
-            present(alert, animated: true, completion: nil)
-            return
+            let vc = INUIEditVoiceShortcutViewController(voiceShortcut: wrapper.original)
+            vc.delegate = self
+            editedIndexPath = indexPath
+            present(vc, animated: true, completion: nil)
+            
+        case .addShortcut:
+            addShortcut()
         }
-        perform(segue: StoryboardSegue.Shortcuts.connectToSegueIdentifier)
     }
     
-    private func addEnable() {
-        addShortcut(with: EnableVPNIntent())
-    }
-
-    private func addDisable() {
-        addShortcut(with: DisableVPNIntent())
+    // MARK: ShortcutsIntentDelegate
+    
+    func shortcutsDidSelectIntent(intent: INIntent) {
+        pendingShortcut = INShortcut(intent: intent)
+        navigationController?.popToViewController(self, animated: true)
     }
     
-    private func addTrustWiFi() {
-        addShortcut(with: TrustCurrentNetworkIntent())
-    }
+    // MARK: INUIAddVoiceShortcutViewControllerDelegate
     
-    private func addUntrustWiFi() {
-        addShortcut(with: UntrustCurrentNetworkIntent())
-    }
-    
-    private func addTrustCellular() {
-        addShortcut(with: TrustCellularNetworkIntent())
-    }
-    
-    private func addUntrustCellular() {
-        addShortcut(with: UntrustCellularNetworkIntent())
-    }
-    
-    private func addShortcut(with intent: INIntent) {
-        guard let shortcut = INShortcut(intent: intent) else {
-            return
-        }
-        let vc = INUIAddVoiceShortcutViewController(shortcut: shortcut)
-        vc.delegate = self
-        present(vc, animated: true, completion: nil)
-    }
-    
-    @IBAction private func close() {
-        dismiss(animated: true, completion: nil)
-    }
-}
-
-@available(iOS 12, *)
-extension ShortcutsViewController: INUIAddVoiceShortcutViewControllerDelegate {
     func addVoiceShortcutViewController(_ controller: INUIAddVoiceShortcutViewController, didFinishWith voiceShortcut: INVoiceShortcut?, error: Error?) {
+        guard let voiceShortcut = voiceShortcut else {
+            dismiss(animated: true, completion: nil)
+            return
+        }
+
+        wrappers?.append(ShortcutWrapper.from(voiceShortcut))
+        wrappers?.sort()
+        reloadModel()
         tableView.reloadData()
+        
         dismiss(animated: true, completion: nil)
     }
     
     func addVoiceShortcutViewControllerDidCancel(_ controller: INUIAddVoiceShortcutViewController) {
+        dismiss(animated: true, completion: nil)
+    }
+
+    // MARK: INUIEditVoiceShortcutViewControllerDelegate
+    
+    func editVoiceShortcutViewController(_ controller: INUIEditVoiceShortcutViewController, didUpdate voiceShortcut: INVoiceShortcut?, error: Error?) {
+        guard let indexPath = editedIndexPath, let voiceShortcut = voiceShortcut else {
+            return
+        }
+        editedIndexPath = nil
+        wrappers?[indexPath.row] = ShortcutWrapper.from(voiceShortcut)
+        wrappers?.sort()
+        tableView.reloadData()
+
+        dismiss(animated: true)
+    }
+
+    func editVoiceShortcutViewController(_ controller: INUIEditVoiceShortcutViewController, didDeleteVoiceShortcutWithIdentifier deletedVoiceShortcutIdentifier: UUID) {
+        guard let indexPath = editedIndexPath else {
+            return
+        }
+        editedIndexPath = nil
+        wrappers?.remove(at: indexPath.row)
+        reloadModel()
+
+        dismiss(animated: true) {
+            self.tableView.deleteRows(at: [indexPath], with: .automatic)
+        }
+    }
+
+    func editVoiceShortcutViewControllerDidCancel(_ controller: INUIEditVoiceShortcutViewController) {
+        editedIndexPath = nil
         dismiss(animated: true, completion: nil)
     }
 }